增加tlibc

+++ tlibc/source/protocol/tlibc_abstract_reader.c
@@ -97,8 +97,7 @@ static tlibc_error_code_t add_name(tlibc_abstract_reader_t *self, const char *na
 {
 	if(self->enable_name)
 	{
-		int len = snprintf(self->name_ptr, TLIBC_READER_NAME_LENGTH - (self->name_ptr - self->name)
-			, ".%s", name);
+		int len = snprintf(self->name_ptr, (size_t)(TLIBC_READER_NAME_LENGTH - (self->name_ptr - self->name)) , ".%s", name);
 
 		if(len < 0)
 		{
@@ -116,8 +115,7 @@ static tlibc_error_code_t del_name(tlibc_abstract_reader_t *self, const char *na
 	if(self->enable_name)
 	{	
 		char curr_name[TLIBC_READER_NAME_LENGTH];	
-		int len = snprintf(curr_name, TLIBC_READER_NAME_LENGTH - (self->name_ptr - self->name)
-			, ".%s", name);
+		int len = snprintf(curr_name, (size_t)(TLIBC_READER_NAME_LENGTH - (self->name_ptr - self->name)), ".%s", name);
 
 		if(len < 0)
 		{
@@ -195,8 +193,7 @@ tlibc_error_code_t tlibc_read_vector_element_begin(tlibc_abstract_reader_t *self
 {
 	if(self->enable_name)
 	{
-		int len = snprintf(self->name_ptr, TLIBC_READER_NAME_LENGTH - (self->name_ptr - self->name)
-				, "[%u]", index);
+		int len = snprintf(self->name_ptr, (size_t)(TLIBC_READER_NAME_LENGTH - (self->name_ptr - self->name)), "[%u]", index);
 
 		if(len < 0)
 		{
